Layouts: Tutorial 14 Video 2:30 in depth description.
A page is a UI that is unique to a route.

A layout is a UI that is shared between multiple pages in the app.

Creating a layout is defined by exporting a React comp from a layout.tsx file.

Root layout is a manditory layout, is at same level as the home page nested right under app.