Routing Meta Data.
Next's Metadata api allows us to define metadata for each page to ensure accurate and relevant information is displayed when pages are shared or indexed.

Configuring meta data: export static metadata object or export a dynamic generateMetadata funciton.

Metadata Rules:
Both layout and page files can export metadata. If it is defined in a layout it applies to all pages in that layout. However, if defined in a page it applies only to that page. Metadata is read in order from the root level down to the final page level.

When there is meta data in multiple plages for the same route they get combined. The deepest one takes priority if any merge conflicts.

HOWEVER, page metadata will replace layout metadata if they have same properties.